+++ redsea-back/redsea-database/DTOs/Banking/ClientGeneralFinances.cs
@@ -1,6 +1,6 @@
 ﻿namespace redsea_database.DTOs.Banking;
 
 public record ClientGeneralFinances(
-    Guid ClientId, 
-    float Balance,
-    float CertificatesBalance);
+    Guid ClientId,
+    decimal Balance,
+    decimal CertificatesBalance);
